requirements.txt
# Build dependencies
setuptools>=60.0.0
wheel>=0.37.0

# Protocol Buffers
protobuf>=6.31.1
grpcio-tools>=1.75.0
pycryptodome>=3.20.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Production dependencies (add as needed)
# requests>=2.28.0
# numpy>=1.21.0